+++ aim/web/ui/src/components/Table/Table.tsx
@@ -11,7 +11,7 @@ import BusyLoaderWrapper from 'components/BusyLoaderWrapper/BusyLoaderWrapper';
 import ResizeModeActions from 'components/ResizeModeActions/ResizeModeActions';
 import ErrorBoundary from 'components/ErrorBoundary/ErrorBoundary';
 
-import { rowCeilSizeConfig, RowHeightSize } from 'config/table/tableConfigs';
+import { ROW_CELL_SIZE_CONFIG, RowHeightSize } from 'config/table/tableConfigs';
 
 import useResizeObserver from 'hooks/window/useResizeObserver';
 
@@ -79,6 +79,7 @@ const Table = React.forwardRef(function Table(
     deleteRuns,
     hideSystemMetrics,
     className = '',
+    focusedState,
     ...props
   }: ITableProps,
   ref,
@@ -112,6 +113,16 @@ const Table = React.forwardRef(function Table(
   });
 
   let groups = !Array.isArray(rowData);
+
+  React.useEffect(() => {
+    if (focusedState && !focusedState.active) {
+      activeRowKey.current = null;
+    }
+  }, [focusedState]);
+
+  React.useEffect(() => {
+    updateFocusedRow(`rowKey-${activeRowKey.current}`);
+  }, [selectedRows]);
 
   React.useImperativeHandle(ref, () => ({
     updateData: updateData,
@@ -234,7 +245,7 @@ const Table = React.forwardRef(function Table(
         } else {
           activeRowKey.current = rowKey;
         }
-        updateHoveredRow(`rowKey-${activeRowKey.current}`);
+        updateFocusedRow(`rowKey-${activeRowKey.current}`);
       } else {
         tableRef.current?.setActiveRow(rowKey);
       }
@@ -248,9 +259,18 @@ const Table = React.forwardRef(function Table(
           const rowCell = document.querySelector(
             `.Table__cell.rowKey-${rowKey}`,
           );
-
           if (!!rowCell) {
-            const top = rowCell.offsetTop - (groups ? 3 : 2) * rowHeight;
+            let top = 0;
+            if (groups) {
+              rowCell.parentElement?.childNodes?.forEach((item, index) => {
+                if ([...item.classList].includes(`rowKey-${rowKey}`)) {
+                  top =
+                    rowCell.parentElement?.offsetTop + rowHeight * (index - 3);
+                }
+              });
+            } else {
+              top = rowCell.offsetTop - 2 * rowHeight;
+            }
             if (
               tableContainerRef.current.scrollTop > top ||
               tableContainerRef.current.scrollTop +
@@ -281,7 +301,7 @@ const Table = React.forwardRef(function Table(
                 // TODO: probably need useEffect for this
                 setTimeout(() => {
                   window.requestAnimationFrame(() => {
-                    updateHoveredRow(`rowKey-${rowKey}`);
+                    updateFocusedRow(`rowKey-${rowKey}`);
                     scrollToElement();
                   });
                 }, 100);
@@ -373,7 +393,7 @@ const Table = React.forwardRef(function Table(
       offsetHeight: tableContainerRef.current.offsetHeight,
       scrollHeight: tableContainerRef.current.scrollHeight,
       itemHeight: rowHeight,
-      groupMargin: rowCeilSizeConfig[rowHeight].groupMargin,
+      groupMargin: ROW_CELL_SIZE_CONFIG[rowHeight].groupMargin,
     });
 
     startIndex.current = windowEdges.startIndex;
@@ -399,7 +419,7 @@ const Table = React.forwardRef(function Table(
         activeRowKey.current = row.key;
       }
 
-      updateHoveredRow(`rowKey-${activeRowKey.current}`);
+      updateFocusedRow(`rowKey-${activeRowKey.current}`);
     }
 
     if (typeof onRowClick === 'function') {
@@ -410,11 +430,17 @@ const Table = React.forwardRef(function Table(
   }
 
   function updateHoveredRow(activeRowClass) {
+    const prevActiveRow = document.querySelectorAll('.Table__cell.focused');
+    if (!!prevActiveRow && prevActiveRow.length > 0) {
+      prevActiveRow.forEach((cell) => cell.classList.remove('focused'));
+    }
     if (activeRowClass !== 'rowKey-null') {
       window.requestAnimationFrame(() => {
-        const prevActiveRow = document.querySelectorAll('.Table__cell.active');
-        if (!!prevActiveRow && prevActiveRow.length > 0) {
-          prevActiveRow.forEach((cell) => cell.classList.remove('active'));
+        const prevHoveredRow = document.querySelectorAll(
+          '.Table__cell.hovered',
+        );
+        if (!!prevHoveredRow && prevHoveredRow.length > 0) {
+          prevHoveredRow.forEach((cell) => cell.classList.remove('hovered'));
         }
 
         const activeRow = document.querySelectorAll(
@@ -422,7 +448,30 @@ const Table = React.forwardRef(function Table(
         );
 
         if (!!activeRow && activeRow.length > 0) {
-          activeRow.forEach((cell) => cell.classList.add('active'));
+          activeRow.forEach((cell) => cell.classList.add('hovered'));
+        }
+      });
+    }
+  }
+
+  function updateFocusedRow(activeRowClass) {
+    const prevHoveredRow = document.querySelectorAll('.Table__cell.hovered');
+    if (!!prevHoveredRow && prevHoveredRow.length > 0) {
+      prevHoveredRow.forEach((cell) => cell.classList.remove('hovered'));
+    }
+    if (activeRowClass !== 'rowKey-null') {
+      window.requestAnimationFrame(() => {
+        const prevActiveRow = document.querySelectorAll('.Table__cell.focused');
+        if (!!prevActiveRow && prevActiveRow.length > 0) {
+          prevActiveRow.forEach((cell) => cell.classList.remove('focused'));
+        }
+
+        const activeRow = document.querySelectorAll(
+          `.Table__cell.${activeRowClass}`,
+        );
+
+        if (!!activeRow && activeRow.length > 0) {
+          activeRow.forEach((cell) => cell.classList.add('focused'));
         }
       });
     }
@@ -454,7 +503,7 @@ const Table = React.forwardRef(function Table(
         offsetHeight: tableContainerRef.current.offsetHeight,
         scrollHeight: tableContainerRef.current.scrollHeight,
         itemHeight: rowHeight,
-        groupMargin: rowCeilSizeConfig[rowHeight].groupMargin,
+        groupMargin: ROW_CELL_SIZE_CONFIG[rowHeight].groupMargin,
       });
 
       startIndex.current = windowEdges.startIndex;
@@ -468,7 +517,7 @@ const Table = React.forwardRef(function Table(
           offsetHeight: target.offsetHeight,
           scrollHeight: target.scrollHeight,
           itemHeight: rowHeight,
-          groupMargin: rowCeilSizeConfig[rowHeight].groupMargin,
+          groupMargin: ROW_CELL_SIZE_CONFIG[rowHeight].groupMargin,
         });
 
         startIndex.current = windowEdges.startIndex;
@@ -511,11 +560,15 @@ const Table = React.forwardRef(function Table(
   React.useEffect(() => {
     if (custom) {
       requestAnimationFrame(() => {
-        updateHoveredRow(
-          `rowKey-${
-            activeRowKey.current ? activeRowKey.current : hoveredRowKey.current
-          }`,
-        );
+        if (!activeRowKey.current) {
+          updateHoveredRow(
+            `rowKey-${
+              activeRowKey.current
+                ? activeRowKey.current
+                : hoveredRowKey.current
+            }`,
+          );
+        }
       });
     }
   }, [custom, listWindow]);
@@ -933,6 +986,10 @@ function propsComparator(
   }
 
   if (prevProps.hiddenColumns !== nextProps.hiddenColumns) {
+    return false;
+  }
+
+  if (prevProps.focusedState !== nextProps.focusedState) {
     return false;
   }