+++ aim/web/ui/src/pages/Metrics/Metrics.tsx
@@ -233,6 +233,7 @@ function Metrics(
                   onRowSelect={props.onRowSelect}
                   archiveRuns={props.archiveRuns}
                   deleteRuns={props.deleteRuns}
+                  focusedState={props.focusedState}
                 />
               ) : null}
             </BusyLoaderWrapper>