+++ aim/web/ui/src/types/components/Table/Table.d.ts
@@ -62,6 +62,7 @@ export interface ITableProps {
   archiveRuns?: (ids: string[], archived: boolean) => void;
   deleteRuns?: (ids: string[]) => void;
   className?: string;
+  focusedState?: any;
 }
 
 export interface ITableRef {